git-svn-id: https://dolphin-emu.googlecode.com/svn/branches/texcache-rewrite@7532 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/FifoPlayer/FifoDataFile.h
@@ -79,7 +79,7 @@ public:
 
 	void AddFrame(const FifoFrameInfo &frameInfo);
 	const FifoFrameInfo &GetFrame(int frame) const { return m_Frames[frame]; }
-	int GetFrameCount() { return m_Frames.size(); }
+	int GetFrameCount() { return (int)m_Frames.size(); }
 
 	bool Save(const char *filename);
